pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Tatiana_Garbuzova <tatiannna0697@gmail.com>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.1"
python-dotenv = "^1.0.1"
dj-database-url = "^2.1.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.9"
whitenoise = {extras = ["brotli"], version = "^6.6.0"}
django-bootstrap5 = "^23.4"
django-filter = "^23.5"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
coverage = "^7.4.1"
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
